Update examples and move them to separate packages
